Refactor

+++ WordleTest2/applicationTest.cpp
@@ -21,7 +21,7 @@ namespace WordleTest2
 		{
 			string truth = "ABCDE";
 			string word = "AXXXX";
-			string pattern = PatternToStringOfSquares(ComputePattern(word, truth), word.size());
+			string pattern = PatternToStringOfSquares(ComputePattern(word, truth), (int)word.size());
 			Assert::AreEqual((string)"VGGGG", pattern); // 🟩⬛⬛⬛⬛ one good
 		}
 
@@ -29,7 +29,7 @@ namespace WordleTest2
 		{
 			string truth = "ABCDE";
 			string word = "XAXXX";
-			string pattern = PatternToStringOfSquares(ComputePattern(word, truth), word.size());
+			string pattern = PatternToStringOfSquares(ComputePattern(word, truth), (int)word.size());
 			Assert::AreEqual((string)"GJGGG", pattern); // ⬛🟨⬛⬛⬛ one misplaced
 		}
 
@@ -37,7 +37,7 @@ namespace WordleTest2
 		{
 			string truth = "ABCDE";
 			string word = "AEXXX";
-			string pattern = PatternToStringOfSquares(ComputePattern(word, truth), word.size());
+			string pattern = PatternToStringOfSquares(ComputePattern(word, truth), (int)word.size());
 			Assert::AreEqual((string)"VJGGG", pattern); // 🟩🟨⬛⬛⬛ one good one misplaced
 		}
 
@@ -45,7 +45,7 @@ namespace WordleTest2
 		{
 			string truth = "ABCDE";
 			string word = "AAXXX";
-			string pattern = PatternToStringOfSquares(ComputePattern(word, truth), word.size());
+			string pattern = PatternToStringOfSquares(ComputePattern(word, truth), (int)word.size());
 			Assert::AreEqual((string)"VGGGG", pattern); // 🟩⬛⬛⬛⬛ one good once
 		}
 
@@ -53,7 +53,7 @@ namespace WordleTest2
 		{
 			string truth = "ABCDE";
 			string word = "XAAXX";
-			string pattern = PatternToStringOfSquares(ComputePattern(word, truth), word.size());
+			string pattern = PatternToStringOfSquares(ComputePattern(word, truth), (int)word.size());
 			Assert::AreEqual((string)"GJGGG", pattern); // ⬛🟨⬛⬛⬛ double misplacement of the same letter
 		}
 
@@ -61,7 +61,7 @@ namespace WordleTest2
 		{
 			string truth = "AABCD";
 			string word = "AXAXX";
-			string pattern = PatternToStringOfSquares(ComputePattern(word, truth), word.size());
+			string pattern = PatternToStringOfSquares(ComputePattern(word, truth), (int)word.size());
 			Assert::AreEqual((string)"VGJGG", pattern); // 🟩⬛🟨⬛⬛ same letter twice one good, a second copy misplaced
 		}
 
@@ -69,7 +69,7 @@ namespace WordleTest2
 		{
 			string truth = "AABCD";
 			string word = "AAXXX";
-			string pattern = PatternToStringOfSquares(ComputePattern(word, truth), word.size());
+			string pattern = PatternToStringOfSquares(ComputePattern(word, truth), (int)word.size());
 			Assert::AreEqual((string)"VVGGG", pattern); // 🟩🟩⬛⬛⬛
 		}
 
@@ -77,7 +77,7 @@ namespace WordleTest2
 		{
 			string truth = "AABCD";
 			string word = "AAXXA";
-			string pattern = PatternToStringOfSquares(ComputePattern(word, truth), word.size());
+			string pattern = PatternToStringOfSquares(ComputePattern(word, truth), (int)word.size());
 			Assert::AreEqual((string)"VVGGG", pattern); // 🟩🟩⬛⬛⬛
 		}
 
@@ -85,7 +85,7 @@ namespace WordleTest2
 		{
 			string truth = "AAACD";
 			string word = "AAXXA";
-			string pattern = PatternToStringOfSquares(ComputePattern(word, truth), word.size());
+			string pattern = PatternToStringOfSquares(ComputePattern(word, truth), (int)word.size());
 			Assert::AreEqual((string)"VVGGJ", pattern); // 🟩🟩⬛⬛🟨 there is a third copy somewhere
 		}
